+++ packages/browser-wallet/src/popup/pages/IdentityIssuance/i18n/en.ts
@@ -1,7 +1,7 @@
 const t = {
     title: 'Your Concordium identity',
     abortExplanation:
-        'The identityissuance was aborted. If you did not abort the process, please try again, or contact support.',
+        'The identity issuance was aborted. If you did not abort the process, please try again, or contact support.',
     successExplanation:
         'Your request has been submitted to the identity provider. It may take a little while for them to confirm your identity. \n\n Once your identity has been verified, you will be able to open an account with it.',
     errorExplanation: 'The identity issuance failed due to: "{{reason}}"',